Fix failed commit of SplitRequestsFailedException move
