+++ java/statestore/src/main/java/sleeper/statestore/dynamodb/FailedDynamoDBSplitRequests.java
@@ -22,11 +22,11 @@ import com.amazonaws.services.dynamodbv2.model.TransactionCanceledException;
 
 import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.SplitFileReferenceRequest;
-import sleeper.core.statestore.SplitRequestsFailedException;
 import sleeper.core.statestore.exception.FileNotFoundException;
 import sleeper.core.statestore.exception.FileReferenceAlreadyExistsException;
 import sleeper.core.statestore.exception.FileReferenceAssignedToJobException;
 import sleeper.core.statestore.exception.FileReferenceNotFoundException;
+import sleeper.core.statestore.exception.SplitRequestsFailedException;
 
 import java.util.ArrayList;
 import java.util.List;
